refactor(conformance): name disconnect probe module
